Plotting now available
As you can see from multiple files, it took me a number of tries
to start in the right direction. I was inundating myself with class
and variable hierarchy and started to write out defining without a
clear roadmap. What I needed was starting with simplest functionality
first. Then I started to see which classes fit under where and what
additional variables and methods were needed. ChatGPT was and will
continue to be a great help when figuring out the holes in my knowledge.